Map the default background color to normal intensity
This is consistent with the default foreground color, and it makes the
default value for the default background color black instead of dark gray.

While theoretically we could use bright colors for background, it makes
more sense to use normal colors for the background as they are darker,
which is more readable. This also is how VGA text modes works, see

http://en.wikipedia.org/wiki/VGA-compatible_text_mode

In that mode the foreground color index has 4 bits while the background
color index has 3 bits, i.e. a range of 0 to 7, which maps to the normal
intensity colors.